+++ tests/models.py
@@ -1,4 +1,15 @@
+from django.contrib.contenttypes.fields import GenericRelation
 from django.db import models
+
+
+class LogEntry(models.Model):
+    content_type = models.ForeignKey(
+        to="contenttypes.ContentType",
+        on_delete=models.CASCADE,
+        related_name="+",
+    )
+    object_pk = models.CharField(max_length=255)
+    event = models.CharField(max_length=100)
 
 
 class Group(models.Model):
@@ -15,6 +26,9 @@ class Widget(models.Model):
     value = models.PositiveIntegerField(default=0)
     other = models.CharField(max_length=100, null=True)
     owner = models.ForeignKey(Owner, null=True, on_delete=models.SET_NULL)
+    logs = GenericRelation(
+        LogEntry, content_type_field="content_type", object_id_field="object_pk"
+    )
 
 
 class Thing(models.Model):